Run helm test with custom ServiceAccount
This allows a Role to be assigned to a test being run
so that it can query services in order to find endpoints
to verify.  Now we don't need to know exactly what the
couchbase cluster name will be.

Also adds a testID key, because when running 'ct' it
wasn't clear exactly which test was failing.

Change-Id: If151f465c5dd4d5fe7e0371d1d1b2f625b62db66
Reviewed-on: http://review.couchbase.org/106014
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>