Simplify renaming of initial bucket
This change puts bucket name inline so that 'default' bucket
is just the synonym for initial bucket, but it's actual name is an attribute
so we can do things like:

 --set couchbaseCluster.buckets.default.name=travel-sample

Btw, buckets are maps because of the  way helm does overrides via --set and -f.
Maps make it easy to access and change specific bucket attrs.

Change-Id: Ibb9f087b4c2a047ff3c2ac8d009feca3a289f211
Reviewed-on: http://review.couchbase.org/106146
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>