Don't require claims to be bound
We should just trust the storageclass to work however it's configured,
rather Immediate or Lazy.  At a minimum the Claim should be in pending state
which is enough to proceed in waiting for Pod to start, and if that succeeds then it
means everything went ok, otherwise we could easily see mount failures in Pod logs.

From config perspective, this removes the need to specify annotations
thee claim relating to underlying storageclass.

Change-Id: I50d6661be6eb866364496403e31d5dff7448bef8
Reviewed-on: http://review.couchbase.org/106476
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>