K8S-915: Fix Log Retention Once And For All
These are so damned flaky!  General spring cleaning then updates to the
generic validators that handle race conditions more gracefully.

I got to the bottom of the non-determinism and vanishing events.
K8S-919 raised for a proper fix.  For now the only solution is to stop
checking the event streams.

Change-Id: I1254e29ecb80c42b70abed6ecb5de0e2e6ad6b47
Reviewed-on: http://review.couchbase.org/106854
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>