K8S-955: CRD Version 2
Create the new V2 CRD with decomposed buckets for starters.  Updated
everything to start using the V2 API, with the exception of the
admission controller which needs to look after both revisions during
upgrade and all the bucket types.  Unit testing the validation was
becoming a nightmare, so we just use QE.

Change-Id: I55735db7f31a1f0b802ef9b7f23bd6fbd14f8b6b
Reviewed-on: http://review.couchbase.org/110240
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>