K8S-953: Library Upgrade
Chances are we will need to go to 1.13 to support versioned CRDs so lets
get this out of the way now as it's a big task.  Additionaly we need to
upgrade the operator-sdk and start using the controller-runtime.

I've chosen to drop automatic creation of the CRD as it seems like an
uncommon option, most people use Openshift and all our testing installs
the CRD separately.  The runtime ability to alter the logging level is
also dead as it must be set by the zap flags, just restart the operator.

Also gone are the readiness nonsense that's redundant and punters kept
moaning about.

Change-Id: I9bd215632f9b8929cc30a6b62349526d958fa27e
Reviewed-on: http://review.couchbase.org/109046
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>