K8S-956: Zap Logging
Rids ourself of logrus, which is an important steeping stone to
providing debug logging of the couchbase client code.  It also removes a
whole bunch of hacks trying to propagate the pointer around.

As a bonus (accidentally commited into this) the entire runloop nonsense
is discarded in favour of using the native requeuing feature of the
controller runtime.  And finally the output when showing spec diffs was
hard to read, I've replaced it with an actual diff!

Change-Id: Ie29f2adee55646fa1e25236a3bbc738e366f5464
Reviewed-on: http://review.couchbase.org/109315
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>