K8S-1006: CRD Reorg
At the moment most things exist at the tyop level of the CRD spec, this
means no logical grouping of related options which makes analysis and
documentation somewhat of a challenge.  This adds security, networking
and logging sections.  Also renames ServerSettings and AuthSecret to be
either more descriptive or terse.

Change-Id: Ic90325bbc4b127c391fa1085eb5f420977fbd88a
Reviewed-on: http://review.couchbase.org/111029
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>