K8S-1010: Explicit CRD Schema Validation
GKE 1.13, even though the data structures support versioned CRD
validation, does not support it.  It needs a control plane flag to
enable it.  So rather than rely on the underlying platform to do it
implicitly we just do it ourselves at the correct point.

One issue was that conversion from a versioned validation to a generic
one caused a crash as enumerations don't work.  Seeing as they are being
removed in structural schema/CRD auto generation we may as well replace
them with patterns now rather than get bitten in the future.

Change-Id: I8ee33a82124f276d9d5411293079e684e0cdd0dc
Reviewed-on: http://review.couchbase.org/111144
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>