K8S-1013: Cleanup Status Handling
We keep a copy of the status in memory and perform updates on that,
which is kinda confusing in that it's part of the main CouchbaseCluster
resource.  By doing the logically correct thing we can catch better
where we aren't updating the status properly.  In my opinion we should
update everything each time, not rely on local caching.  This is
important for when the status isn't preserved, e.g. Velero backups.

Change-Id: I1cebc9b581ed04f8c1407db1f0528eded5d4d367
Reviewed-on: http://review.couchbase.org/111865
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>