K8S-1032: QE Zen
The admin console stuff is totally redundant now, so save the characters
by sending it to hell.  So too (finally!) is the old way of doing
events, burn in hell with your friend.  DeletePod, burn in hell.  Burn,
burn, burn.  Cleaned up the RZA stuff, because why use 20 lines when you
can use 2.

Change-Id: I589750ded3e6351f09976e6de8bfe232e1613215
Reviewed-on: http://review.couchbase.org/112219
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>