K8S-1035: QE FIxes 22/07/19
Operator test was racy, so added in some more synchronization.  Failover
test doesn't actually work.  6.0.0 just plows on, and doing an rm -rf of
the directory isn't going to work as the directly is repopulated before
the directory unlink happens.  PV tests can take a very long time to
spin up new nodes so increase the timeouts.  Status test, is racy, there
is only a small window of time the status is fully empty before the
operator marks it as paused, which does stay stable, so use that.

Change-Id: I67696aafd05ed7f61bf6c6843d1d3991bbedc537
Reviewed-on: http://review.couchbase.org/112281
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>