K8S-1012: Perform Unstructured JSON Mutation
The API pipeline goes, defaulting -> validation.  We perform defaulting
by decoding before schema validation and thus are not being model
citizens, introducing potentially proprietary error messages.  This
performs defaulting (mutation) on unstructured JSON so no decoding is
required.  In doing so I've tighened up the data structures so that
fields are unspecified in the JSON when not specified in the YAML, also,
especially for buckets, we were being too strict, certain fields can be
safely defaulted and not forced to be required (e.g. zero values are ok
and implicit e.g. 0 and false.

Change-Id: Ib6ca66c4dfe220224dad3ee4b924c4ef4e90b888
Reviewed-on: http://review.couchbase.org/112580
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>