K8S-788: Update Unready Service Mechanism
The old way of doing things has been deprecated since 1.11.  We need to
remove this and use the new parameter when upgrading services.  As a
temporary hack I've added to the update functions, we should probably
move to the same method of doing things as pod upgrades eventually e.g.
kubectl apply style.

Change-Id: Ib2558126ff4f010a9b7a4a272e88d4392ccd181c
Reviewed-on: http://review.couchbase.org/116931
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>