K8S-1056: Test Fixes
TLS tests were suffering from a timeing change, so failover did not
happen within the 2 minute limit, extended to 5.  Autofailver test
wasn't working because the bucket was still in warmup when we killed, so
added a barrier in there.  Saw a timeout on EKS while changing service
types, so increased the timeout, and also added collections to services
so we gather any associated events.

Change-Id: I5b4001ce56f49347fa6fde5be223abb6e3b856ec
Reviewed-on: http://review.couchbase.org/113074
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>