K8S-1059: Clean Up RetryOnErr
It's a clear observation that this function, and its child are only
really used with context timeouts, and the number of retries is utterly
superflous (as are the stringy based parameters that used to be logged).
Long term goal finally realized!  While exploring error propagation I
discovered I had inadventently broken delta recovery failures.  The most
striking point is that there was zero code coverage so I've added in a
test, which was quite hard to trigger in all honesty, so not a bad bug
and could be worked around.

Change-Id: Ia406681107491b00e7ef44a5ebfe80d3b803052a
Reviewed-on: http://review.couchbase.org/113257
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>