K8S-1063: Disable Readiness
We do our readiness checks with pods/exec in order to constrain "ready"
to mean data availabilty is guaraneteed in the event of a pod death due
to eviction (provided you have 1 replica that is).  Some customers don't
allow this, and are happy to run an unsafe database, so we provide the
means to disable readiness entirely until we can come up with a HTTP
based solution -- a sidecar most likely.

Change-Id: I96e0f1c29e6cd0ba2c1a972421adbd238d128963
Reviewed-on: http://review.couchbase.org/113625
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>