K8S-1067: Fuzzy Deployment Collection
Users normally don't read the documents so when performing log
collection on OpenShift they don't add in the correct operator image
flag.  We default to couchbase/operator:2.0.0 however needs to be
"registry.connect.redhat.com/couchbase/operator:2.0.0-1" on OpenShift
systems.  By doing a substring match we have a better chance of working
on all platforms.  Some may say doing away with the version would be
better still, but tools are tied to the release, so we have an excuse.

Change-Id: I53a46522d751941e97b7256aebd8423482568dbe
Reviewed-on: http://review.couchbase.org/113898
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>