K8S-1094: Operator Restart Races
In a lot of tests we kill the operator to see what happens across
restart, or when things are modified across restarts.  This leads to
race conditions depending on how on-the-ball the Deployment controller
is.  We should just manually take down the deployment and recreate it
giving us far more control over the exact synchronization of operations.

Change-Id: Iefd4fae9a7bfef510c572dffc8d63efefb7da36c
Reviewed-on: http://review.couchbase.org/115402
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>