K8S-1095: Full Pod Upgrade Path
At present we look at the pod's annotated version when performing an
upgrade.  Wouldn't it be nice to look at anything that was set when
originally creating the pod and perform an upgrade?  Think retroactive
application of resource limits etc.  The upgrade mechanism gives us a
way to upgrade pods (which are typically immutable) in a Couchbase safe
way, extend it to cover any change.  We encode the pod configuration as
a JSON encoded annotation (like kubectl apply) as we canont use the raw
pod configuration due to defaulting, and use this to detect changes.

Change-Id: Ibed1ccc582f0b2d48d19f5ef23ee7dd303821ace
Reviewed-on: http://review.couchbase.org/116435
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>