K8S-1140: API Caching
Every ~5 seconds we poll pods, service, etc in order to check whether
the intended view on the world matches reality.  However do this too
much and the Kubernetes API risks getting flooded with requests.  As a
good citizen we should aspire to perform only what is necessary in order
to not be responsible for customer downtime/support escalation.  The
notion of shared informers simply caches resource by doing a list, then
a watch to keep the cache in sync.

Change-Id: I04e2ed81624abfbce489c4c20ce130439422320d
Reviewed-on: http://review.couchbase.org/117658
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>