K8S-1234: Better Rebalancing
With 6.5.0-beta2 we discovered that rebalance may not be instantly visible,
with 6.5.0-rc1 we discovered they may never be seen at all!  Make your
rebalance, trigger, be observed, report.  If it isn't observed then
check whether the cluster status is as expected (implying it happened)
then get on with your life.

Change-Id: Ifc4017b85f16db5cabedf78d1e0168f2583991bf
Reviewed-on: http://review.couchbase.org/119022
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>