K8S-1191: Upgrade Documents
Separate the upgrade docs into concepts and how-tos.  For the concepts
I've added in rules concerning upgrade paths, pod upgrades and volume
upgrades.  Perhaps breaking the rules a little bit, but I put in an
adminition we can refer to when users complain it's taking too long so
we can blame server instead!

Change-Id: I6cbcbb5f0b351cd006a1d9ab4b89e81dd1e0ad2a
Reviewed-on: http://review.couchbase.org/119954
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>