K8S-1237: Server Group Docs
Split this into a high level concept and a low-level how to.  Also take
the opportunity to hopefully clarify things and the docs should be
specific and not leave anything to the imagination, like my old version
did at times.  I've also removed waffle that didn't have any relevence
to keep the focus.

Change-Id: I102a559e3f3ec823d05a0a0bb67fbc926d9d5188
Reviewed-on: http://review.couchbase.org/120040
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>