K8S-1296: Refactor /spec/servers/0/pod
We gat a lot of can we override this, can we override that in the pod
specifications, and each time around we have to do some work, which adds
in delays.  What is slowly happening is that this attribute is turning
into PodTemplateSpec, so bite the bullet and just do it now to support
anything we wish to modify in the future.

Change-Id: I10d2ba9098e39f9bc1dc405de1617e63a581babe
Reviewed-on: http://review.couchbase.org/121191
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>