K8S-1268: Reorganize Docs
There is a clear pattern emerging, so we should go all in while we have
the chance (e.g. SEO and stuff to worry about).  By naming docs as
concepts, howtos, reference and tutorials we are setting clear
expectations about what level the content should be and its audience.
Links were all patched with a global sed command, so should be
programatically correct for little physical effort.

Deleted items are now either replaced or are just the standard so we
don't need to reiterate or confuse matters with repetition.

Change-Id: If83ea713e29ffc6c790aaee9dc07220c39e51282
Reviewed-on: http://review.couchbase.org/121403
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>