K8S-1311: Make QE Namespace Configurable
Almost all of the code is "reentrant", it uses the namespace of an
object, with the exception of all infrastructure creation and deletion.
The plague of Global spreading like a virus infecting everything.  In
order to test DNS based remote networking we need to run in different
namespaces, not a single one, so make this configurable per cluster,
with the exeception of the DAC which we keep hard coded to default.
This allows multiple virtual clusters to map to the same physical, and
also accounts for QE's special snowflakes who will still have instances
running on them that need to be deleted/upgraded per run.

Change-Id: I8fcdecb5aab0c72388b5fb2d0c61f867753ffab2
Reviewed-on: http://review.couchbase.org/121569
Reviewed-by: Tommie McAfee <tommie@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>