K8S-1294: Docs Versioning
We have a bunch of references to the operator, server and prometheus
versions, these need updating.  It makes life a huge amount simpler if
we just update once and everywhere, so make Eric overjoyed and partials
and attributes for this.  Also now we are aware of attributes, fix any
warnings that cropped up in the Anotra output...

Change-Id: I9461511e77c4a7c76b5f441ba730d467e91dc719
Reviewed-on: http://review.couchbase.org/122128
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Simon Murray <simon.murray@couchbase.com>