K8S-940: Revisit Cloud
We got a bit too excited with cloud.  We shouldn't tell people how to
use AKS, EKS or GKE, it's not our problem to support those who have the
time and money to do so themselves.  I've retained the majority as a
tutorial as there is some important stuff there regarding peering.
Things that are directly relevant to getting started have been added to
a cloud pre-reqs.

Change-Id: Ifc465df3d501e0bda8afb1efe4d92049b041db4d
Reviewed-on: http://review.couchbase.org/122263
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>