K8S-1352 Backup backoffLimit
Currently we only have a hard-coded BackOffLimit (the amount of times a
job will retry before failing) for the restore job.

Make this configurable for the restore job and also for backup jobs
spawned by the Cronjobs. Current backoffLimit for backup jobs is 6 which
is way too high.

AKA if a backup job fails to execute Completed by the backoffLimit we
wait until the next job is spawned by the Cronjob.

Apply a default value of backoffLimit of 2 (so maximum 3 failed pods per
job).

Change-Id: I956fa11198ceded7c45d6d8781edbc189ad1efa9
Reviewed-on: http://review.couchbase.org/122609
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>