Make the Makefile the Source for Versioning
It stands to reason that the build manifests determine the version
number for packaging, so they should also determine the version number
embedded in all the tools.  I've extended the tools to all have the
option of a version string so we can know exactly what people are
running.  More pertinently, while git stuff is not available, we do have
the build number, so put that in to prevent the wrong stuff being run.
Finally I've added in the concept of a revision (e.g. beta1) which is
again driven by the Makefile, so while it isn't used yet, it can be in
the future.

Change-Id: If51bb0305f05608d8e45a99accf6b0d2cf66eb98
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/124367
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Hillery <ceej@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>