K8S-1321: Remove Management Library
The dictionary defintion of a hindrance.  This first one merges gocbmgr
into couchbaseutil, and begins to make QE and the main product
independant.  The key observation here is if we share code too much then
we may do something bad, causing them to do something bad and possibly
masking a problem, the adding foregin node stuff is an example of bad.

Change-Id: I582a0a9b4913a6ac467d2f283ffd323ac3b2b86f
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/127787
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>