K8S-1474: Refactor API Accesses
This turn the API stuff into a simple "call couchbase" library.  Any
retry logic can then be summarized as call API x for nodes y.  The goal
here is to provide a set of canonical API call functions rather than a
whole file of wrappers as exists at the moment.

Change-Id: I48f73779458687342e2f95ec05e7b09409a9c26b
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/128212
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>