K8S-1482: Improve API Typing
Couchbase has the concept of an OTP node and a host name, neither of
which are intuitive to grok or use.  By attaching types to these
entities we can better handle their flow around the system and limit all
the translation that goes on between different key spaces.  The goal, as
always, simpler easier and safer code.  As an added bonus, faster too,
as the OTP node stuff can be inferred (it's impossible to have multiple
nodes on the same pod).

Change-Id: If1c0a83001662e8430516fec289c9f7e525fe941
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/128477
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>