Pipelines For QA
Why have twenty jobs when you can have one?  This allows the testing of
any gerrit commit at any point in the review lifecycle.  It does not
allow the live hacking of configuration: all configuration that is run
is part of change control.  All credentials are handled securely.

The test creates ephemeral images from the source tree, rather than
relying on the build team, so you can test real code before commit.
Also I found it slightly crazy to have to create a file to run a test,
when you can just pass in the parameters on the CLI.  This is now doing
half of what tco does, why it wasn't done this way to start with is
beyond me!

Change-Id: I9466a90706a363337043905dc383334eed6b506f
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/129539
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Prateek Kumar <prateekkumar276@gmail.com>