K8S-1492: Skip readiness flagging if already ready
Motivation
----------
Under certain platforms, at least GKE and AKS, this readiness
check has shown to cause a memory leak in underlying libraries.
Note that this is not happening on all platforms, EKS for
instance.

Modifications
-------------
Change FlagPodReady to skip doing any work if a given container
is already showing a ready status.

Results
-------
No leak because we effectively rate limit how often we call the
API that was holding memory.

Change-Id: I173c692689ef1ff7b9f186d7021a96720ee1bbc9
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/129735
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>