K8S-1041: Sort Out Error Handling - Part I
There are a lot of different ways of doing errors, which is inconsistent
for starters.  Second we want to use golangci-lint to force us to use
wrapped errors, which allows errors to be unwrapped, inspected and the
root cause divined.  This has the nice property of very strictly
defining where the operator injects errors into the system, at which
point we can attach context in the form of backtraces that allow us to
better reason about what happened when a potential defect is hinted at.
Obviously the hard bit is also identifying where errors come from a 3rd
party library, but more on that in a bit...

Change-Id: I635b6c4d9becdd77bdedccef388cc11b10e02221
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/131264
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>