K8S-1519: Support Multiple Image Pull Secrets
The official build process has changed docker registry providers... for
half of the images we need to test, which means we need multiple pull
secrets. Ostensibly it's easy, just use an array instead of a variable,
however... We now generate the secret names and clean them up with a
label selector.  This means that the non-deterministic number of secrets
also differ per-namespace.  This is actually a good thing as we can now
cache those secrets in the cluster object and propagate them to where
they are needed, freeing us from global variables.  Additionally the
operator deployment cannot be cached globally because the configuration
changes based on namespace.

Change-Id: I1c86855527373116b61f7198b8093d2697836ddb
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/130219
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Prateek Kumar <prateekkumar276@gmail.com>