K8S-809: Remove Hard Coded Cluster Names
The configuration is overly complex and not ever used.  It's just a
list of tests that use the number of clusters they require.  Mapping to
a named cluster is mostly pointless when they are homogeneous.
Furthermore this now opens up the possibility of testing concurrently!
Few little tweaks to code organization, stuff was in totally the wrong
place.  I've also added optional colour for those of use who respond to
green/red things and highlighting in our peripheral vision, it's a much
nicer formatting than the default!

Change-Id: Ib52f4f8286f6cad75b24e5a70c964dda7216688f
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/130312
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Prateek Kumar <prateekkumar276@gmail.com>