K8S-1522: Enable Individual Test Runs
Allow '-test' to be specified on the CLI to override the default
behaviour of running a full suite, which finally makes the framework
usable to a human being without having to maintain 100s of lines of
python scripting.  Cluster names are meaningless now, so I've taken the
liberty of removing them, which makes the configuration even cleaner
now.

Change-Id: Ica9c664ffd543f9932df188e11be3d485aff29cf
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/130400
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>