Improve Reporting Accuracy
We must gather information about the test within the test.  From a
parent function we can only see whether it passed, we also need to know
whether it crashed or whether it was skipped.  This does all of that and
actually emits correct(er) junit XML.  I've deleted retries as they just
mask problems.  I'd rather strive for 100% reliability rather than fake
it.

Change-Id: Ib2c3e2c597784ffd103dcf3f1cf4896e1cfae908
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/130532
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Prateek Kumar <prateekkumar276@gmail.com>