K8S-1536: Expect Add Nodes in Rebalance
No one understands node reconcile, what is a known node anyway?  We know
all about each and evey node from a variety of sources!!  Anyway, during
PVC recovery 2 come up, fail over the third, which happens to recover in
time to go into add back, which is not considered "known" as regards
rebalance is concerned, and thus not expected to be part of the cluster,
which it is.  I'm not going to do what I want to do, as it's 4pm on a
Friday...

Change-Id: Ifa6a2418913e741917cccbe9f6029b10c6758798
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/130981
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>