K8S-114: Admin Password Rotation Part I
This sets the ground work.  Our view of the password is ephemeral, and
can change from that of Couchbase server.  As such we need to store our
idea of the last working password, that's our source of truth.  The
cluster persistent storage needs to be a secret now, so provide an
upgrade path.  Finally do the rotation if necessary, raise and event and
check it works with a test.

Change-Id: I8b6d0149e63d260faac5cf41a79a17a566b20d83
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/131382
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>