K8S-1168: Non-Exec Node Readiness
This bins the exec check (note it will cause an upgrade to existing
clusters, nothing can be done about this), but means we have now
banished pods/exec from the Operator for all customers.  Seeing as there
were no tests for this GA feature (!!), I've taken the liberty of adding
some.  The minimum Kubernetes version is now 1.14+.

Change-Id: I8453d28d227dff5cc7881af3c559f175db3f5ff6
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/131644
Reviewed-by: Simon Murray <simon.murray@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>