K8S-1567: Ensure PVC Deletion
PVCs are a source of truth as far as members ae concerned, previously we
relied on cached state, which is bad.  Now we poll reality there is a
gap between a PVC being deleted and it not being there.  Wait for
deletion and also ignore deleted volumes, nothing else we can do within
reason to ensure things are behaving as we expect.

Change-Id: I36bf9cf7d8f7496058ebb1cb12405f4e6a27e214
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/131940
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>