K8S-1568: Remove Reliance On Resource Status
For a start things will go crazy if you backed up and restored a cluster
in mid upgrade as the status isn't preserved, so it must go in the
cluster secret.  I've taken the liberty to back out my own
overengineering of the upgrade/rollback logic because it's hacky and
bring little real value.  Kinda relevant is the fact that "upgrade"
isn't strictly an upgrade any more, so versions are irrelevant in 90% of
upgrade cases.

Change-Id: Iba897cb2b827e1e851c8ca0b1afcfcb3c4f8fbee
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/132021
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>