K8S-1599: Fix Test Regressions
As expected, when we lowered the failover time out, some events don't
get seen (in particular nodes failing over before going down).  This
isn't a bad thing, the operator is working as designed and it's a test
race condition.  There is a race in the new SDK tests, ensure that the
bucket is alive before continuing.  We need to raise an event when pod
creation is rejected synchronously, this must be new Kubernetes
behaviour.  The RZA failover test incorrectly deleted volumes, when it
shouldn't have.  They are needed for recovery!

Change-Id: I502a90f12c13ba57201805ee9226ce66c2e4a157
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/133435
Reviewed-by: Prateek Kumar <prateekkumar276@gmail.com>
Tested-by: Build Bot <build@couchbase.com>