K8S-1218: Hibernation Part I
Basic hibernation that just shoots the cattle in the head, and makes
sure it can recover.  Waiting for all queues to drain _is_ going to be a
pain in the arse, I just know.  Technically it may be cleaner to scale
down "statless" server classes, but that makes life *much* harder than
force failover and recovery.

Change-Id: Ib98f9abda843e046c3f11138b6c233d96ecb6f8d
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/133811
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>