K8S-1611: QE Fixes
The weekly QE stability fixes... Scheduling tests actually alter on the
fly, especially with multiple tests concurrently as nodes go in and out
of NoSchedule, so start to make these tests more tolerant, forcing them
to always fail, and handling the event steams altering based on temporal
behaviour.  Ensure the LDAP service is running before letting Couchbase
use it, in the hope that CBS is broken and we'll not see bad behaviour
again.  Finally improve debug logging to make problem determination more
streamlined.  There is a race in the PVC recovery test, so it does
different things at different times, so handle that.

Change-Id: Ibe17fb0354fdc42029ab99bf656ad87f0bc13cc1
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/134982
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Prateek Kumar <prateekkumar276@gmail.com>