K8S-1680: Finish cbopcfg Updates
This was left in a half completed state, this finishes it off.  We use
cobra for comnmand parsing so it works exactly like all other Kubernetes
tools.  An example command is 'cbopcfg generate operator' so we can in
future have a 'cbopcfg create operator' for example.  We could even have
it do CRD installation in a future release, even incorporate cbopinfo
into it e.g. 'cbopcfg collect logs`...

Change-Id: I6adb389e78387d2aa38ee277ede7fd53f3349058
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/136626
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Roo Thorp <roo.thorp@couchbase.com>