K8S-1499: Add ability to specify storageclass for backup
Currently the backups created by the Operator just use the default
storage class.
This is far too restrictive in a multi-tenant Kubernetes cluster where
the user might not have direct control over what storage class is
offered as the default.

We need a way of specifying the storage class to be used by the backup
in the CouchbaseBackup resource.

Change-Id: Ifd9c893ffdc5bedce227791b0abff626b6ddfbd1
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/135985
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>