K8S-1572: Speed Up Validation
The critical path is the negative validation for creation, where we
check things that don't validate tell the user about it.  Split this up
into component specific sections which is more conducive to concurrent
speediness.  Also, for a long time it's bugged me we have to create,
delete and recreate the operator, all of which is utterly pointless, so
allow each test to be configured to do/not do things as defined by the
test.

Change-Id: I75fc594c66d2145bf8fcbc3646b0afb6b0b72433
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/139281
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Roo Thorp <roo.thorp@couchbase.com>