Improve Validation Testing
Make validation a required pass.  More importantly make it test all the
possible Kubernetes versions because that is where the incompatibilities
lie--at the API layer.  It also gives us a one-up on QE by having most
of the certification problems solved ahead of time.

Due to the parallel nature of the tests we need to discriminate between
suites on different platforms, so add in a flag for that.  I also
discovered that 1.15 is broken, so fix that up (hard coded error strings
again).

Change-Id: Iad85fc0d93169f06a115e2084572fb3ec1f1cc98
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/139470
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Roo Thorp <roo.thorp@couchbase.com>