K8S-1697: Istio/Networking Docs
Define the tiny Istio specific hack.  However, I noticed that everything
else networking was out of date, so I've updated all the references with
the new way, flagged the old way as deprecated, and ported all the
examples over to how we want people to do things.  Remarkably little
because of DRY documentation!

Change-Id: I5e05981d0b9ddddca342416cf29a6e13d5eea583
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/140232
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Ingenthron <ingenthr@gmail.com>