K8S-1783: Upgrade Testing
Finally, I'm kicking this annoyance in the head.  Would it be better to
juect checkout and compile the old code, probably, is it easier, maybe
not as every build is done differently.  Work in progress certainly, and
most importantly doesn't actually run multiple versions at the same
time, however that's just a case of upgrading the CRDs and DAC while
leaving the operator in place.

Change-Id: Ib25b19703bb3f67a22e5ee82f0bf60b4f5b5a936
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/140902
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Prateek Kumar <prateekkumar276@gmail.com>