K8S-1802: Auto Resource Allocation
In general, this is a good idea as it makes testing in parallel more
stable from a contention perspective, it also passes that benefit on to
customers.  The bad side is that it will cause yet another upgrade, and
may stop people's clusters from working... to be fair they didn't
configure it correctly in the first place!

Change-Id: I8d24c0e6c8b4a4f9210d8aa92b2c973ce1b812aa
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/141144
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>