K8S-1596: Update Linter
Bring in all the new improvments and bug fixes for code enforcement.
And as per usual it's pointed out shite code, mostly to do with error
wrapping, or the lack thereof, but it also pointed out Retry was awful.
I've been meaning to kill it for some time, none like the present!

Change-Id: I337f16d760fd9e4bcc883f265d99e6baf877048c
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/141376
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>