K8S-1812: Implicit Upgrade Fix
We used to require this in order to apply new defaults to existing
clusters, but somehow the DAC is now invoked to do it in later versions.
Long story short, when restarting the Operator it attempts to fill in
any missing defaults, but it tries applying JSON patches to a struct and
fails badly.

Change-Id: Ic7f764e60593350b46a4b6590c6e14334a7a711b
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/141419
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>