K8S-1167: Mutable Server Groups
Really quite simple actually, the strip scheduler now keeps track of
where valid things are, but also where invalid things are.  We then use
an equal-cost graph traversal algorithm to find the shortest path from
where our state is, where it should be.  This is mead eaiser by the fact
that upgrades and topology changes are mutually exclusive, so we don't
need to track the scheduling state changes for now.

Change-Id: I89a363857a25ff3314c31ee9f85134cba18fe23c
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/143179
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tommie McAfee <tommie@couchbase.com>