K8S-1824: Pod Reconcile
The label/annotation reconcile code assumes that topology is in a good
state i.e. upgrades have been done, before running.  However if an
upgrade request comes in when the cluster is in warm up, then topology
is silently ignored, and the pod upgrade stuff happens, overwriting the
pod spec annotation and thus making the operator think the upgrade has
happened, when it hasn't.  This preserves the important annotations for
now, and I've added a note that the behaviour needs some additional
thought.

Change-Id: I231827943493173cb081e8c2512ecb2c93ec8194
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/143257
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Ma <daniel.ma@couchbase.com>