K8S-1963: Password Rotation Race
We know it's possible for the password to be rotated and the operator to
be taken down before the cache is updated (thus locking ourselves out),
but we need the tests to be stable.  This waits for the password updated
event to happen so we know everything is in a sane state before
terminating the operator.

Change-Id: I78b6a0eefa96efb8cbbf96349a6d1bb24e6f7df4
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/144358
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Roo Thorp <roo.thorp@couchbase.com>