K8S-1801: More DAC Configuration Options
There are two facets to this.  First, fuddie duddies who don't like us
reading secrets in order to validate they are correctly formatted and
valid (not that we would ever let this be a security risk).  Second
those who use 3rd party tooling that creates clusters before secrets and
the like.

Change-Id: I1af836b3b3035f36161d2fec914b112e31af7b92
Reviewed-on: http://review.couchbase.org/c/couchbase-operator/+/144711
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pat <patrick.stephens@couchbase.com>
Reviewed-by: Simon Murray <simon.murray@couchbase.com>